os-system-back/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.6.0
pymongo==4.9.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1
pydantic-settings==2.1.0